chore: move test files to be colocated
