+++ src/routes/raids/index.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
-import { loadingScreen, render, screen, userEvent } from '../tests/testUtils'
-import App from '../App'
-import { fetchedCollectionData } from '../tests/data/raidsData'
+import { loadingScreen, render, screen, userEvent } from '../../tests/testUtils'
+import App from '../../App'
+import { fetchedCollectionData } from '../../tests/data/raidsData'
 
 beforeAll(() => {
   jest.doMock('../utils/useCollection', () =>
